Fix issues with querying using labels
Issue was that when we get the label in the query, e.g. "Room 123", it
technically has a blank namespace because its stored as a goraptor.URI.
The Brick.ttl file defines a blank abbreviation "" as mapping to the
actual Brick namespace, which confused the query execution. This fix
stops us from adding an abbreviation that's the blank namespace. No
additional tests fail as of writing, so hopefully this is a sufficient
fix