Intersect var re-definitions
Should fix #28, but uncovered some other issues with operating on
non-snapshots of the database. This will require some internal
restructuring, but should help with some of the race conditions we're
hitting on occasionally failing queries with OR clauses. This commit
also adds some small pauses to the tests to help those pass