Run queries off of a database snapshot
This is most of the fixes necessary to make sure that queries run off of
a snapshot of the database rather than the live version (this has been a
long time coming). This introduces a new 'snapshot' type to perform
these operations and traverse the multiple leveldb instances under the
covers.